+++ packages/worphling/src/cli/Cli.ts
@@ -11,12 +11,18 @@ export class Cli {
     private detectFlags(): Flags {
         const args = minimist(process.argv.slice(2), {
             boolean: ["try-exact-length", "with-sorting"],
-            default: { "try-exact-length": false, "with-sorting": false },
+            default: {
+                "try-exact-length": false,
+                "with-sorting": false,
+            },
         });
 
         const isTryingExactLengthEnabled = args["try-exact-length"];
         const isSortingEnabled = args["with-sorting"];
 
-        return { isTryingExactLengthEnabled, isSortingEnabled };
+        return {
+            isTryingExactLengthEnabled,
+            isSortingEnabled,
+        };
     }
 }